BlogArticle: fix editor crashing (#2648)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes


![Screenshot 2023-06-29 at 16.47.22.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/07c7cfc9-02ad-4ee9-97e0-aa9104680388/Screenshot%202023-06-29%20at%2016.47.22.png)


Prevent Storyblok editor from crashing when editing blog articles

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

React crashes when editing since categories are returned as a list of UUIDs instead of being resolved as Story objects.

I show a placeholder when this happens.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code