Blog: fetch all article teasers (#2653)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Fetch all article teasers instead of first 25

Render ingress as paragraph tag

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

I make the build fail if we have more than the max number of articles in one locale. This will be our trigger to implement pagination but if we never reach this limit we don't need to implement pagination.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code