GEN-652 | Make sure price intent is synced to current product (#2677)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Update Price Intent on page navigation

- Check if completed Price Intent belongs to the current product

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

The price intent is not synced with the current product when navigating to a new page.

Instead you see the price of the previous product after you navigate.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code