Move Checkbox from UI to Onboarding (#2680)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Move Checkbox component from UI kit to Onboarding app

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

The component is only used in Onboarding app, so it makes sense to move it there.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code