Render 404 if slug page is not found (#2699)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Correctly render 404 page when a CMS page is not found.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Now we just throw an exception which causes 500 error.

Perhaps this will allow us to use "fallback: blocking" to autopublish new pages!

## Checklist before requesting a review

- [ ] I have performed a self-review of my code