GEN-666 | CartItem: make the whole card clickable to expand (#2702)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

https://github.com/HedvigInsurance/racoon/assets/1220232/9eeeabb5-8deb-4a30-a1c6-8ef8ea16be24

- Make the "whole" Cart Item card clickable to exapand and collapse the details

- Add hover state to the Cart Item card

- Make the action buttons medium size

- Make the full collapsible "header" clickable

😈

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

I decided to only make the top part of the card clickable instead of litteraly the whole card. This is to avoid overlapping with the action buttons. Click events propagate so if the whole card is clickable, it will be triggered even if the user clicks on the action buttons e.g.

Instead of trying to work around this, I decided to only make the top part clickable.

I used the `:has` parent selector to only apply the hover state when hovering over the top part of the card. This has surprisingly good browser support already at 87%.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code