Add Slack endpoint to trigger a deploy of the website (#2720)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Allow a Slack user to deploy the website again

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

This can be useful after you published a new page in Storyblok for example

## Checklist before requesting a review

- [ ] I have performed a self-review of my code