GEN-748 | Add CRM Retargeting Landing Page (#2741)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

![Screenshot 2023-07-11 at 11.50.36.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/eb737c5f-f85b-40c4-bbf5-cadf87bced1a/Screenshot%202023-07-11%20at%2011.50.36.png)

- Implement a draft version of the CRM landing page to be used for re-targeting

- List all price intents that have been confirmed but not yet added to cart

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Design is just a draft but this is necessary to understand if we need to make any changes to the overall idea and/or API.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code