Add resume shop session form to debug dialog (#2801)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes


![Screenshot 2023-07-17 at 16.08.13.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/ffa486c3-8171-4869-9780-37f99c6a6028/Screenshot%202023-07-17%20at%2016.08.13.png)


- Add option to resume a shop session from debug dialog

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Pretty common action

Styles are a bit off but I address that in the next PR

## Checklist before requesting a review

- [ ] I have performed a self-review of my code