Revert apollo client to stable version
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Revert to stable version of apollo client

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Caused issues with how onComplete is called when price intent is updated

## Checklist before requesting a review

- [ ] I have performed a self-review of my code