chore(eslint): enable consistent-type-definitions rule
## Describe your changes

* Enables typescript eslint [_consistent-type-definitions_](https://typescript-eslint.io/rules/consistent-type-definitions) rule

> TypeScript provides two common ways to define an object type: interface and type.
> The two are generally very similar, and can often be used interchangeably. Using the same type declaration style consistently helps with code readability.

Here I'm deciding upon using `type`. With that I had to update the code to disable some edge cases where we need to use `interface`. We could change it to enforce the usage of `interface`,  but that would cause a bigger change since we've been using `type` in our codebase for the majority of the cases.

## Justify why they are needed

This is part of a series of PRs that includes new/updated rules added by typescript-eslint v6. The idea is to use those PRs so we discuss as a team if we want to enable a particular rule or not.