chore(eslint): enable prefer-nullish-coalescing rule
## Describe your changes

* Enable typescript eslint [_prefer-nullish-coalescing_](https://typescript-eslint.io/rules/prefer-nullish-coalescing) rule.

> The `??` nullish coalescing runtime operator allows providing a default value when dealing with `null` or `undefined`. Because the nullish coalescing operator only coalesces when the original value is `null` or `undefined`, it is much safer than relying upon logical OR operator chaining `||`, which coalesces on any falsy value.

I've configure it to ignore conditional cases and `string` and `boolean` primitives.

## Justify why they are needed

This is part of a series of PRs that includes new/updated rules added by typescript-eslint v6. The idea is to use those PRs so we discuss as a team if we want to enable a particular rule or not.