chore(eslint): document why we want to disable no-redundent-type-constituents-rule
## Describe your changes

Document why we want to disable typescript eslint [_no-redundant-type-constituents_](https://typescript-eslint.io/rules/no-redundant-type-constituents) rule

## Justify why they are needed

This is part of a series of PRs that includes new/updated rules added by typescript-eslint v6. The idea is to use those PRs so we discuss as a team if we want to enable a particular rule or not.