Trials debugger: set start date to today (#2934)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Set the start date in Trials debugger to today

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Otherwise we don't see the expected Trials extension information when navigating to the Widget

## Checklist before requesting a review

- [ ] I have performed a self-review of my code