Add grid decorator to Storybook
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes
- Move decorators to separate file
- Add new gridDecorator

I thought it would be neat to better match Figma and Storybook stories in terms of layout. For example, showing Accordions as full screen is a bit misguiding since we will always restrict the width when implemented in the app. The only option we have today is the `layout` parameter that has `centered` or  `fullscreen`. I've added the GridLayout in some story templates but that feels like too much boilerplate when writing stories.

Instead we can have it as a global decorator that you enable by setting a `grid` parameter in the stories you want to use it.
```js
parameters: {
  grid: '1/2' or { base: '1', md: '5/6', lg: '2/3', xl: '1/2' },
}, 
```

### Before
<img width="1728" alt="Screenshot 2023-08-11 at 15 14 11" src="https://github.com/HedvigInsurance/racoon/assets/6661511/64842453-1063-49c9-a4c3-cd094641559a">

### After
<img width="1725" alt="Screenshot 2023-08-11 at 15 42 18" src="https://github.com/HedvigInsurance/racoon/assets/6661511/9362291d-f18f-4293-8d12-ec7e04a10554">


<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed
Make story layouts match Figma/implementation

## Checklist before requesting a review

- [ ] I have performed a self-review of my code