Only automerge during business hours (#2967)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Restrict renovate automerge to business hours so that we can quickly fix problems if they appear

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code