Add OneTrust cookie list styling
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes
Add OneTrust cookie list styling. Makes it easier to change to override it in the component than having it in OneTrust dashboard, so I migrated it over from there. I kept it in string format to make migration back and forth easier ( and it's a pain to write class name styling with object syntax 😆)


https://github.com/HedvigInsurance/racoon/assets/6661511/f82251a1-5fb9-4ddc-af6c-46762ad52bb9


<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code