Rennovate: ignore adyen SDK (#3079)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Ignore updates to Adyen SDK in Rennovate

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We are not able to upgrade to the latest version of Adyen SDK

## Checklist before requesting a review

- [ ] I have performed a self-review of my code