Log context when shop session is not found (#3214)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Log cookies when we are not able to find the current shop session (checkout).

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

This happens surprisingly often but not sure why so trying to collect more context around the issue.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code