Raise Renovate rate limit
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Raise hourly limit on Renovate PR creation.  This seems to be a bottleneck in current setup.  See comment in config for more details

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code