refact(productReviews): validate average ratings/review comments data
## Describe your changes

- Update `productReviews` service by validate average rating / review comments data, logging an error when they are invalid;
- Makes `ProductReviewCommentsBlock` simplier by removing some safe verifications as we've increased the the trust on getting data in the right format.

## Justify why they are needed

After some discussions we thought that would make `ProductReviewCommentsBlock` simplier. Another approach would be to make sure nothing fish get's written into the database (vercel/kv), but I guess we can start with this as it's easier to implement and also easier to let us now when something is wrong with the data.