Add new PriceBreakdown component for car dealership (#3524)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes
A new component for explaining the cost for trial and extension separately.  
I didn't want to use the `Price` component where we send in a reduced price. The logic is the other way around in this case. The trial price is the fixed price (`amount`) and the default price that we want to strike through is the price for the same tier in full price (`defaultAmount`). 

<img width="986" alt="Screenshot 2023-11-08 at 11 53 11" src="https://github.com/HedvigInsurance/racoon/assets/6661511/d0320c11-1a2a-4c15-b0c4-61346f2d151e">


<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed
It doesn't make sense to show a total price for explaining the price the member will pay for 2 months trial and then the extension

## Checklist before requesting a review

- [ ] I have performed a self-review of my code