Track active insurance details tab as query param (#3536)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Implement Robin's idea that active tab in insurance details should be reflected in URL

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

So that user would stay on the same tab after reloading the page

## Checklist before requesting a review

- [ ] I have performed a self-review of my code