CAR-37 - fix(ConfirmationPage): show total only when cart is not empty
## Describe your changes

- Show `<TotalAmmountContainer>` only for confirmation pages about shop sessions that actually has something included into the chat.

That would avoid this issue:

![image](https://github.com/HedvigInsurance/racoon/assets/19200662/30371fee-0145-49f9-bf92-9cf94fae032f)

## Justify why they are needed

I know this is a workaround. Been talking with William and ideally we want confirmation page as a totaly CMS driven page. If we get that it would be just about not adding that "TotalBlock" for Car Dealership confirmation pages. Until we're not there, I thought that could be enough.