Refactor Widget flow start route (#3594)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

This refactoring breaks down the large function into smaller, more manageable functions, each with a single responsibility.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

It makes the code easier to understand and maintain.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code