Update new-member redirect to insurance page (#3650)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes
See title
<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed
Request from Peter. We have quite a few backlinks so it would be better to point it to the insurance page instead. 
## Checklist before requesting a review

- [ ] I have performed a self-review of my code