Add support for line breaks in Lokalise (#3680)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes
Make sure line breaks in Lokalise are rendered to better control text output. This is useable for texts that should have the same line break for all viewport sizes. If they will differ, we should adjust it with balance and width settings instead.

### Before
<img width="400" alt="Screenshot 2023-11-29 at 17 08 12" src="https://github.com/HedvigInsurance/racoon/assets/6661511/73c09bc9-91a8-43d4-a492-d4955a811ac4">

### After
<img width="391" alt="Screenshot 2023-11-29 at 17 07 47" src="https://github.com/HedvigInsurance/racoon/assets/6661511/a8708305-3352-4517-853a-c01f17a4037e">

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed
Better control of setting text in the app
## Checklist before requesting a review

- [ ] I have performed a self-review of my code