feat(Tooltip): increase tapable area on mobile
## Describe your changes

* Tooltip: increase trigger tap area for coarse pointers (e.g: finger)

The solution is presented in a CSS which I can't link here. However it was inspired by this [codepen](https://codepen.io/third774/pen/XWgXZRY)

**Desktop**

<img width="568" alt="Screenshot 2023-12-04 at 07 30 10" src="https://github.com/HedvigInsurance/racoon/assets/19200662/2a469873-7eba-4c06-9ff4-39913a430563">

**Mobile**

<img width="381" alt="Screenshot 2023-12-04 at 07 20 29" src="https://github.com/HedvigInsurance/racoon/assets/19200662/7e909a0f-25e1-4bfa-9419-d297dfaabd37">

## Justify why they are needed

Makes it more "accessible" for mobile users since it's hard to be precise while clicking on such small button as the tooltip (16px x 16px). Got this trick from a CSS course. Apple recommends at tapbable areas of at least 44px x 44px.

Ideally we wan't that included into design system, as it might influence on the designs.