Don't allow picking a start date in the past (#3718)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Set a minimim start of today in the start date input.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

I feel like we discussed this before but it seems very reasonable

## Checklist before requesting a review

- [ ] I have performed a self-review of my code