GEN-1528 / Implement bundle title and pillows for QuickAdd component (#3726)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes
Show "Home + Accident" title and double pillows when we have a home insurance in the cart. I realised that we would still show the QuickAdd component if you remove your home from the cart and have some other product left since I suppose we don't remove the offer recommendation. So in that case, we should display it without home.

### With home in cart
  
<img width="488" alt="Screenshot 2023-12-12 at 15 09 19" src="https://github.com/HedvigInsurance/racoon/assets/6661511/e1a80926-debb-4440-91a1-cd147b8db73e">


### After removing home from cart
<img width="494" alt="Screenshot 2023-12-12 at 15 09 45" src="https://github.com/HedvigInsurance/racoon/assets/6661511/9ffac618-42f2-41e7-8d17-993c297a2768">

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code