GEN-1673 - refact(memberReviews): unify api
## Describe your changes

Now that _trustpilotData_ and _productReviewsData_ belong to the same feature - `memberReviews` - I've created a single API for `memberReviews` that describes what kind of data should be expected while working with reviews. `trustpilot` and `productReviews` data sources then are now compliant with that API which makes the UI simpler as no parsing is need on the client.