fix(hydration-errors): make sure to import useTranslation from next-i18next
## Describe your changes

* Fix hydration errors on _Checkout_ page by changing from where `RemoveActionButton` gets `useTranslation` hook: `next-i18next` instead of `react-i18next`.

https://github.com/HedvigInsurance/racoon/assets/19200662/24f3ec28-ef2c-48d6-a445-e71b5938487c

Also took the liberty of updating it in other places of the codebase as well.
I could say (at lest personally) the major reason why that mistake insists on existing in your codebase is because intelisense only suggest auto importing `useTranslation` from `react-i18next`.  As a consequence I've added a eslint rule to help us to remember.

<img width="1027" alt="image" src="https://github.com/HedvigInsurance/racoon/assets/19200662/66958688-772c-486a-b7d8-1432e313e673">

## Justify why they are needed

I didn't fully understand the issue but here's the context where it happens:

* `i18n` was not being initialised for `react-i18next` during build

In any case we should be using `useTranslation` hook from `next-i18next` so...

That hypothesis can be verified by adding the following to `RemoveActionButton`:

```
  const { t, ready } = useTranslation('cart')
  console.log('ready', ready)
  console.log('t', t('REMOVE_ENTRY_BUTTON'))
```

That would print `false` and `REMOVE_ENTRY_BUTTON` on the server but `true` and `Ta bort` on the client.