Ignore non-ts files for eslint (#3899)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Continue ignoring plain JS files for eslint

One day we should configure our linting in a way that does not require everything to be part of typescript app, but I failed to make it work in a reasonable time

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code