Replace some usages of next/router with next/navigation in main layout (#3929)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

We won't be able to use next/router hooks in app dir, so changed some of the usages related to main layout to next/navigation

next/navigation works both in app and pages dirs

We still need to figure out a replacesment for router.events.on that would work in both routers

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code