fix(TextWithLink): provide a better default for the case no link placeholder was given
## Describe your changes

* Provide a better default for the case no link placeholder was given

## Justify why they are needed

Before we're rendering `<>props.children</>` on such cases and that's probably a mistake. What I think we want is to render `<>{props.children}</>` so `TextWithLink` is still able to render a some text that doesn't include a link placeholder ([[link placeholder]])). I'm also adding a log for such cases as using `Text` is better suitable for such cases.