Add useResponsiveVariant as hydration-safe way to select desktop or mobile (#4021)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Replace most usages of `useBreakpoint` with `useResponsiveVariant`

Main use case if hydration-compatible selection of mobile vs desktop variant

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Main driving goal was to make page header smaller - we don't want to keep desktop menu in DOM on mobile and we don't want to render it either

## Checklist before requesting a review

- [ ] I have performed a self-review of my code