Toggle cookie banner and GTM behind feature flag (#4025)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Only load cookie banner and GTM when feature flag is enabled.  We're going to have it on in production and for debugging when needed and disabled otherwise.  Before it was controlled by GTM

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Do you love cookie banners when debugging?  I don't :)

## Checklist before requesting a review

- [ ] I have performed a self-review of my code