Fix storyblok-cache-version script (#4059)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

We had an error in "get cache version" script which did not break the build, but made it slower.  It also caused a problem with redirects not being updated that I fixed yesterday

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code