GEN-1937 - fix(bankid): launching on IOS
## Describe your changes

Improves a little bit the experience for IOS users that **open up hedvig-com in another browser other than safari**. For the case you've open it on Safari while having another browser as the default one, _Bankid App_ will redirect to your default browser and I'm afraid there's nothing we can do about it. At least the Safari tab that will be on the current state so although not perfect, the user can still go back to hevig-com manually. 

## Justify why they are needed

Raised as an issue during a testing session. We can't solve the actual issue but we can at least make things a little bit better for other cases.