RND-524 - fix(cookie-banner): make sure it doesn't get in the way of our dialogs
# Describe your changes

* Make sure cookie banner consent doesn't get in into the way of our dialogs.

## Justify why they are needed

### The issue

Sometimes, cookie banner get's into the way of our dialogs which can be anoying:

https://github.com/HedvigInsurance/racoon/assets/19200662/9ad982bb-a2d2-420a-bb67-274a7f6ddd08

An easy way to reproduce this is load a product page with price calculator opened on mobile (`?openPriceCalculator=1`) while having disabled the cache.

### The reason

That happens for the cases our dialog gets added into the DOM before the cookie consent. Since both are positioned elements in the same stacking context, DOM order is used to determine should should be on top. In this case, cookie banner will be on top which can be annoying.

<img width="1220" alt="reason" src="https://github.com/HedvigInsurance/racoon/assets/19200662/09520b4c-80b0-4c08-9f22-3b3835b9b81b">


### The solution

[Nextjs recommends](https://shorturl.at/gkAQ8) to have cookie consent scripts loaded with `beforeInteractive` strategy.  If we follow that recommendation we can make sure our dialogs always get added after one trust into the DOM which then will make they override it without having to fight with `z-index` .

https://github.com/HedvigInsurance/racoon/assets/19200662/aef1f352-76db-42bf-9028-81138f8b36a4

As you can see in the video, cookie banner get's show first and then when our app scripts take in, our dialog is shown on top of it.