Storyblok story analysis scripts (#4112)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add scrips for dumping and analyzing CMS stories

For now I need them to deprecate and remove some of page flags, but could be useful in the future as well

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code