Prevent layout shift of cart item while loading desktop menu (#4117)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Prevent cart menu item from jumping on desktop when page loads

It was aligned centrally before this PR and jumped to the right when desktop menu variant appeared

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code