Convert Breadcrumbs styles to vanilla-extract (#4120)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Convert Breadcrumbs styling to vanilla-extract

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

In the next PR I plan to handle `hideBreadcrumbs` page property with pure CSS, which makes it compatible with app router

## Checklist before requesting a review

- [ ] I have performed a self-review of my code