Move bold style from GlobalBanner (emotion) to Banner (vanilla) (#4143)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Move styles for bold text from `Globalbanner` to generic `Banner`

It's useful presentation logic, no reason for it to live in component that's primarily responsible for state management

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code