Reorganize PurchaseForm so that its context providers wrap only form instead of full page (#4214)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Reorganize `PurchaseForm` parent contexts so that they are placed inside as-small-as-possible `Suspense` block.  This fixes search params access in context providers

PriceIntentContext is still a mess and should be eventually refactored to use jotai atom families instead of React state + effects to sync - but that's for the future

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Allows to render most of product page content statically, without layout shift

## Checklist before requesting a review

- [ ] I have performed a self-review of my code