Add bundle discount tooltip to OfferPresenter when priceIntent qualifies for bundle discount (#4277)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Show bundle discount tooltip when presenting a price in price calculator if current product is eligible (not Accident)



<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code