Make shopSession and priceIntent cookie live for 30d (#4328)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

We've found a bug during app router test when going to cart led to using current session along with cart items

Currently our cookies don't specify TTL and can be expired at any time.  Maybe that was what happened.  In any case, having long-lived cookies is what we wanted for a long time for many reasons, so here it is

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code